Extract PRLinkReconciler from IssueTracker (CROW-1094)
Move session↔PR link detection + reconciliation (applySessionPRLinks,
reconcileMissingPRLinks + the build/fetch/apply/resolve helpers, shellSync) and
the reconcile decision statics/types into their own collaborator. It writes
links through the shared, injected JSONStore (ADR 0012 / #728) and appState via
an unowned back-reference, and reaches the shared GitHub error sink through
owner.handleGitHubBackendError (widened to internal, stays with the warnings
concern). refresh() drives applySessionPRLinks / reconcileMissingPRLinks in the
same order.

The class is public so the cross-module session-capability predicates
(canAddMergeLabel / canSetProjectStatus) keep their home; an extension IssueTracker
re-exposes the exact IssueTracker.<symbol> spelling used by AppState resolvers,
RPCHandlers, addMergeLabel's repoSlug parse, and the tests — no call site or
assertion changes.

swift test --package-path Packages/CrowEngine: 789 tests pass.


Crow-Session: 18A4CD55-8B33-48E4-B5EB-E710314DF5E5
