+++ Packages/CrowEngine/Sources/CrowEngine/IssueTracker.swift
@@ -599,6 +599,10 @@ public final class IssueTracker {
     /// Extracted per CROW-1094; `refresh()` still drives it in order.
     lazy var completion = SessionCompletionController(owner: self)
 
+    /// Session ↔ PR link detection + reconciliation, and the public
+    /// session-capability predicates. Extracted per CROW-1094.
+    lazy var reconciler = PRLinkReconciler(owner: self)
+
     public init(appState: AppState, providerManager: ProviderManager, store: JSONStore) {
         self.appState = appState
         self.providerManager = providerManager
@@ -864,7 +868,7 @@ public final class IssueTracker {
         if let ghResult {
             // Session PR link detection runs against open PRs only — we only
             // ever want to attach a fresh link when there's an open PR.
-            applySessionPRLinks(viewerPRs: ghResult.viewerPRs)
+            reconciler.applySessionPRLinks(viewerPRs: ghResult.viewerPRs)
 
             // For sessions with an existing .pr link whose PR isn't in the open
             // viewer set, fetch the state in one batched aliased query. This
@@ -990,7 +994,7 @@ public final class IssueTracker {
         // Runs after the reactive path so we only ask providers for the
         // sessions that actually need it. Safe for GitLab-only or no-GitHub
         // workspaces — the GitHub branch is gated by candidate count.
-        await reconcileMissingPRLinks()
+        await reconciler.reconcileMissingPRLinks()
 
         // Auto-cleanup expired completed/archived sessions. Runs outside
         // the ghResult block so it fires even without GitHub data. Placed
@@ -1325,7 +1329,7 @@ public final class IssueTracker {
     /// IssueTracker UI side-effect (scope warning, rate-limit suspension).
     /// Untyped errors get a console line and otherwise propagate as "this
     /// cycle is degraded" via the caller's nil-return.
-    private func handleGitHubBackendError(_ error: Error, operation: String) {
+    func handleGitHubBackendError(_ error: Error, operation: String) {
         switch error {
         case ProviderError.insufficientScope(let scope):
             reportScopeWarning(scope)
@@ -1593,630 +1597,6 @@ public final class IssueTracker {
     // `AssignedListing` / `MonitoredPRListing` from the backends in
     // `runConsolidatedGitHubQuery` above and consumes them directly.
 
-    // MARK: - Session PR Link Detection (piggyback)
-
-    /// Build an index of viewer PRs keyed by `(repoSlug, branch)` and `url`, then
-    /// attach PR links to sessions whose primary worktree branch matches.
-    private func applySessionPRLinks(viewerPRs: [ViewerPR]) {
-        guard !viewerPRs.isEmpty else { return }
-
-        // Prefer OPEN PRs over closed ones when a branch has multiple.
-        var byBranch: [String: ViewerPR] = [:]  // key = "repo/slug#branch"
-        for pr in viewerPRs {
-            let key = "\(pr.repoNameWithOwner)#\(pr.headRefName)"
-            if let existing = byBranch[key] {
-                if pr.state == "OPEN" && existing.state != "OPEN" {
-                    byBranch[key] = pr
-                }
-            } else {
-                byBranch[key] = pr
-            }
-        }
-
-        // Accumulate new links and persist them in a single store write below.
-        // Writing per-session inside the loop meant N full-store encode + atomic
-        // disk writes when a burst of PRs got linked at once — the dominant
-        // main-thread stall behind the concurrent-review freeze (#304).
-        var newLinks: [SessionLink] = []
-
-        for session in appState.sessions {
-            guard !session.isManager else { continue }
-            let wts = appState.worktrees(for: session.id)
-            let links = appState.links(for: session.id)
-
-            guard !links.contains(where: { $0.linkType == .pr }) else { continue }
-            guard let primaryWt = wts.first(where: { $0.isPrimary }) ?? wts.first else { continue }
-
-            let branch = primaryWt.branch
-            guard !branch.isEmpty else { continue }
-
-            let repoSlug = resolveRepoSlug(worktree: primaryWt)
-            guard !repoSlug.isEmpty else { continue }
-
-            guard let pr = byBranch["\(repoSlug)#\(branch)"] else { continue }
-
-            let link = SessionLink(
-                sessionID: session.id,
-                label: "PR #\(pr.number)",
-                url: pr.url,
-                linkType: .pr
-            )
-            appState.links[session.id, default: []].append(link)
-            newLinks.append(link)
-        }
-
-        guard !newLinks.isEmpty else { return }
-        // Route through the shared, injected `store` — never a throwaway
-        // `JSONStore()`. A fresh instance reads its own (possibly stale) disk
-        // snapshot and its full-store write can silently clobber a session
-        // another writer just added (#728).
-        store.mutate { data in
-            data.links.append(contentsOf: newLinks)
-        }
-    }
-
-    // MARK: - Session PR Link Reconciliation
-
-    /// A session that the reconcile pass should query a provider for. Built from
-    /// non-archived, non-review sessions that have a primary worktree branch
-    /// but no `.pr` link yet.
-    struct ReconcileCandidate: Sendable, Equatable {
-        let sessionID: UUID
-        let provider: Provider
-        let repoSlug: String       // "corveil/crow"
-        let branch: String
-        let gitlabHost: String?    // nil for github.com
-    }
-
-    /// A Jira-tasked session whose PR should be found by the *ticket key* it
-    /// references (e.g. `MAXX-6859`) rather than by branch. Jira PR branches
-    /// are renamed by the working agent and rarely match the session's
-    /// registered worktree branch, so branch matching can't find them.
-    struct ReconcileKeyCandidate: Sendable, Equatable {
-        let sessionID: UUID
-        let provider: Provider     // code provider (.github today)
-        let repoSlug: String
-        let key: String            // "MAXX-6859"
-        let gitlabHost: String?
-    }
-
-    /// A branch match returned by the provider. `state` follows GitHub's
-    /// `PullRequestState` for GitHub and a normalized "OPEN"/"MERGED"/"CLOSED"
-    /// for GitLab (mapping `opened|merged|closed`). `updatedAt` drives
-    /// tie-breaking when a branch has multiple non-OPEN PRs.
-    struct ReconcileBranchMatch: Sendable, Equatable {
-        let sessionID: UUID
-        let number: Int
-        let url: String
-        let state: String
-        let updatedAt: Date?
-    }
-
-    /// Given a set of matches per session, decide which link to create for
-    /// each session. Prefers OPEN over non-OPEN; falls back to most-recent
-    /// `updatedAt`. Deterministic when timestamps are absent (highest `number`
-    /// wins as a stable tie-breaker). Pure — no appState, no I/O.
-    nonisolated static func decideReconcileLinks(
-        matches: [ReconcileBranchMatch]
-    ) -> [ReconcileBranchMatch] {
-        let bySession = Dictionary(grouping: matches, by: { $0.sessionID })
-        var picks: [ReconcileBranchMatch] = []
-        for (_, group) in bySession {
-            guard let pick = group.max(by: { lhs, rhs in
-                // Returns true when lhs should sort BEFORE rhs (i.e. rhs wins).
-                let lhsOpen = lhs.state == "OPEN"
-                let rhsOpen = rhs.state == "OPEN"
-                if lhsOpen != rhsOpen { return !lhsOpen }  // rhs open → rhs wins
-                switch (lhs.updatedAt, rhs.updatedAt) {
-                case let (l?, r?):
-                    if l != r { return l < r }  // newer wins
-                case (nil, _?):
-                    return true                  // rhs has date → rhs wins
-                case (_?, nil):
-                    return false                 // lhs has date → lhs wins
-                case (nil, nil):
-                    break
-                }
-                return lhs.number < rhs.number   // tie-break on number
-            }) else { continue }
-            picks.append(pick)
-        }
-        return picks
-    }
-
-    /// Enforce that a single PR attaches to at most one work item. Groups the
-    /// final per-session picks by PR URL; if a URL is claimed by sessions with
-    /// more than one distinct work-item identity (ticket key, else branch), the
-    /// PR can't be attributed to one of them with confidence, so it is dropped
-    /// from all of them — never guess (#520). Duplicate sessions sharing one
-    /// identity (same key/branch) keep the link. Pure — no appState, no I/O.
-    nonisolated static func dedupeContestedPRs(
-        _ picks: [ReconcileBranchMatch],
-        identityBySession: [UUID: String]
-    ) -> [ReconcileBranchMatch] {
-        let byURL = Dictionary(grouping: picks, by: { $0.url })
-        var out: [ReconcileBranchMatch] = []
-        for (_, group) in byURL {
-            let identities = Set(group.compactMap { identityBySession[$0.sessionID] })
-            if identities.count > 1 { continue }   // contested across tickets → none
-            out.append(contentsOf: group)
-        }
-        return out
-    }
-
-    /// Route a reconcile candidate to a *code* backend. A task-only provider
-    /// (`.jira`) has no code surface, so a session tracked by one resolves PRs
-    /// through its `codeProvider` — mirroring the
-    /// `codeProvider ?? provider` convention in `SessionService.findPRLink` and
-    /// `AutoRespondCoordinator`. Falls back to host sniffing when no
-    /// code-bearing provider is recorded (e.g. sessions predating the field).
-    /// Pure — no appState, no I/O.
-    nonisolated static func resolveReconcileProvider(
-        codeProvider: Provider?, provider: Provider?, host: String
-    ) -> (provider: Provider, gitlabHost: String?) {
-        if let p = codeProvider ?? provider, !p.isTaskOnly {
-            return (p, p == .gitlab ? (host.isEmpty ? nil : host) : nil)
-        }
-        if host == "github.com" || host.isEmpty { return (.github, nil) }
-        return (.gitlab, host)
-    }
-
-    /// Whether `session` may add the `crow:merge` label to its PR — i.e. its
-    /// **code** backend declares `.autoMergeLabel`. Resolves the code provider
-    /// via the `codeProvider ?? provider ?? .github` convention (ADR 0005) so a
-    /// Jira/Corveil-tasked GitHub-code session is gated on GitHub, not on its
-    /// task provider (CROW-532). Pure — easily unit-tested.
-    public nonisolated static func canAddMergeLabel(session: Session, providerManager: ProviderManager) -> Bool {
-        let provider = session.codeProvider ?? session.provider ?? .github
-        return providerManager.codeBackend(for: provider)?.capabilities.contains(.autoMergeLabel) ?? false
-    }
-
-    /// Whether `session` may be moved to a project-board "In Review" status —
-    /// i.e. its **task** backend declares `.projectBoardStatus`. Mirrors the
-    /// retired native `AppState.canSetProjectStatus(for:)` (GitHub Projects v2 /
-    /// Jira: yes; GitLab: no — ADR 0005), which gated the "In Review" button.
-    /// Restores that gate for the web UI (CROW-749). Pure — unit-tested like
-    /// `canAddMergeLabel`.
-    public nonisolated static func canSetProjectStatus(session: Session, providerManager: ProviderManager) -> Bool {
-        guard let provider = session.provider else { return false }
-        return providerManager.taskBackend(for: provider).capabilities.contains(.projectBoardStatus)
-    }
-
-    /// Instance convenience over ``canSetProjectStatus(session:providerManager:)``
-    /// using this tracker's provider manager — the daemon's `list-sessions` gate.
-    public func canSetProjectStatus(for session: Session) -> Bool {
-        Self.canSetProjectStatus(session: session, providerManager: providerManager)
-    }
-
-    /// For each non-archived, non-review session missing a `.pr` link with a
-    /// resolvable (repoSlug, branch), query the provider directly and upsert
-    /// a link when a PR exists on that branch. Runs once per refresh cycle
-    /// after the reactive `applySessionPRLinks` pass.
-    private func reconcileMissingPRLinks() async {
-        let candidates = buildReconcileCandidates()
-        let keyCandidates = buildReconcileKeyCandidates()
-        guard !candidates.isEmpty || !keyCandidates.isEmpty else { return }
-
-        var matches: [ReconcileBranchMatch] = []
-
-        let github = candidates.filter { $0.provider == .github }
-        if !github.isEmpty, let hits = await fetchPRsForReconcile(candidates: github) {
-            matches.append(contentsOf: hits)
-        }
-
-        let gitlab = candidates.filter { $0.provider == .gitlab }
-        let hostsSeen = Set(gitlab.compactMap { $0.gitlabHost })
-        for host in hostsSeen {
-            let forHost = gitlab.filter { $0.gitlabHost == host }
-            matches.append(contentsOf: await fetchGitLabMRsForReconcile(candidates: forHost, host: host))
-        }
-
-        // Jira-tasked sessions: find the PR by the ticket key it references,
-        // since the PR branch won't match the worktree branch. Feeds the same
-        // `decideReconcileLinks` so a key-found and branch-found PR for one
-        // session resolve to a single best pick.
-        matches.append(contentsOf: await fetchPRsByKeyForReconcile(candidates: keyCandidates))
-
-        // Each session's work-item identity (key preferred, else branch) so the
-        // de-dup pass can tell a legitimate duplicate-session match from one PR
-        // being claimed by two different tickets.
-        var identityBySession: [UUID: String] = [:]
-        for c in candidates { identityBySession[c.sessionID] = c.branch }
-        for c in keyCandidates { identityBySession[c.sessionID] = c.key }
-
-        let decided = Self.decideReconcileLinks(matches: matches)
-        applyReconciledPRLinks(Self.dedupeContestedPRs(decided, identityBySession: identityBySession))
-    }
-
-    /// Walk appState and build the set of sessions needing a reconcile pass.
-    /// Runs on MainActor; safe to read appState directly.
-    private func buildReconcileCandidates() -> [ReconcileCandidate] {
-        var out: [ReconcileCandidate] = []
-        for session in appState.sessions {
-            guard !session.isManager else { continue }
-            guard session.status != .archived else { continue }
-            guard session.kind == .work else { continue }  // review sessions get PR links at creation
-            let links = appState.links(for: session.id)
-            guard !links.contains(where: { $0.linkType == .pr }) else { continue }
-
-            let wts = appState.worktrees(for: session.id)
-            guard let primaryWt = wts.first(where: { $0.isPrimary }) ?? wts.first else { continue }
-            guard !primaryWt.branch.isEmpty else { continue }
-
-            let info = resolveRepoInfo(worktree: primaryWt)
-            guard !info.slug.isEmpty else { continue }
-
-            // Route by the *code* provider: a Jira/Corveil task-only session
-            // codes against GitHub/GitLab via `codeProvider`, so resolving on
-            // `session.provider` alone (→ `.jira`) would drop the candidate.
-            // Falls back to host sniffing when no code-bearing provider exists.
-            let (provider, gitlabHost) = Self.resolveReconcileProvider(
-                codeProvider: session.codeProvider,
-                provider: session.provider,
-                host: info.host
-            )
-
-            // GitLab candidates require a known host — GITLAB_HOST env var is
-            // how the glab wrapper picks an auth token. Skip silently rather
-            // than fall through to a wrong-host call.
-            if provider == .gitlab, gitlabHost == nil { continue }
-
-            out.append(ReconcileCandidate(
-                sessionID: session.id,
-                provider: provider,
-                repoSlug: info.slug,
-                branch: primaryWt.branch,
-                gitlabHost: gitlabHost
-            ))
-        }
-        return out
-    }
-
-    /// Build key-based reconcile candidates: Jira-tasked sessions missing a PR
-    /// link, whose PR is discoverable by the ticket key (e.g. `MAXX-6859`)
-    /// rather than by branch. Gated on a Jira ticket URL so GitHub/GitLab-tasked
-    /// sessions are untouched (they keep pure branch matching). Runs on
-    /// MainActor; safe to read appState directly.
-    private func buildReconcileKeyCandidates() -> [ReconcileKeyCandidate] {
-        var out: [ReconcileKeyCandidate] = []
-        for session in appState.sessions {
-            guard !session.isManager else { continue }
-            guard session.status != .archived else { continue }
-            guard session.kind == .work else { continue }
-            let links = appState.links(for: session.id)
-            guard !links.contains(where: { $0.linkType == .pr }) else { continue }
-
-            let wts = appState.worktrees(for: session.id)
-            guard let primaryWt = wts.first(where: { $0.isPrimary }) ?? wts.first else { continue }
-
-            // Resolve the ticket key: prefer a Jira ticket URL, else derive it
-            // from the worktree branch (e.g. `feature/max-monorepo-maxx-7035-…`
-            // → `MAXX-7035`). The branch fallback covers the prefix-drop case
-            // where the PR head loses the repo prefix the worktree carries (#520).
-            //
-            // The branch fallback is gated to task-only trackers (Jira/Corveil):
-            // a lowercased branch can't distinguish a real Jira project ("maxx")
-            // from an ordinary word/repo segment ("api"), so a GitHub/GitLab
-            // issue branch like `feature/acme-api-197-fix` would yield a bogus
-            // "API-197" key. Those sessions resolve via the branch path instead.
-            let urlKey = session.ticketURL.flatMap {
-                Validation.isJiraSpec($0) ? Validation.jiraKey(from: $0) : nil
-            }
-            let branchKey = (session.provider?.isTaskOnly == true)
-                ? Validation.ticketKey(fromBranch: primaryWt.branch) : nil
-            guard let key = urlKey ?? branchKey else { continue }
-
-            let info = resolveRepoInfo(worktree: primaryWt)
-            guard !info.slug.isEmpty else { continue }
-
-            let (provider, gitlabHost) = Self.resolveReconcileProvider(
-                codeProvider: session.codeProvider,
-                provider: session.provider,
-                host: info.host
-            )
-            if provider == .gitlab, gitlabHost == nil { continue }
-
-            out.append(ReconcileKeyCandidate(
-                sessionID: session.id,
-                provider: provider,
-                repoSlug: info.slug,
-                key: key,
-                gitlabHost: gitlabHost
-            ))
-        }
-        return out
-    }
-
-    /// Resolve PR links for Jira-tasked sessions by searching the code repo for
-    /// the ticket key. GitHub only today (the `CodeBackend` default returns no
-    /// matches for providers without text PR search). Best-effort: a backend
-    /// error skips the cycle rather than dropping links.
-    private func fetchPRsByKeyForReconcile(candidates: [ReconcileKeyCandidate]) async -> [ReconcileBranchMatch] {
-        let github = candidates.filter { $0.provider == .github }
-        guard !github.isEmpty, let backend = providerManager.codeBackend(for: .github) else { return [] }
-        do {
-            let matches = try await backend.findPRsMatchingKeys(Self.dedupedKeyCandidates(github))
-            return Self.fanOutKeyMatches(matches, across: github)
-        } catch {
-            handleGitHubBackendError(error, operation: "findPRsMatchingKeys(github)")
-            return []
-        }
-    }
-
-    /// Project `ReconcileKeyCandidate`s onto de-duplicated `(repoSlug, key)`
-    /// pairs for the backend. Mirrors `dedupedBranchCandidates`.
-    nonisolated static func dedupedKeyCandidates(_ candidates: [ReconcileKeyCandidate]) -> [KeyCandidate] {
-        var seen: Set<KeyCandidate> = []
-        var out: [KeyCandidate] = []
-        for c in candidates {
-            let kc = KeyCandidate(repoSlug: c.repoSlug, key: c.key)
-            if seen.insert(kc).inserted { out.append(kc) }
-        }
-        return out
-    }
-
-    /// Fan each `KeyPRMatch` back to every session sharing its `(repoSlug, key)`.
-    /// Mirrors `fanOutMatches` for the branch path.
-    nonisolated static func fanOutKeyMatches(
-        _ matches: [KeyPRMatch],
-        across candidates: [ReconcileKeyCandidate]
-    ) -> [ReconcileBranchMatch] {
-        var sessionsByKey: [KeyCandidate: [UUID]] = [:]
-        for c in candidates {
-            sessionsByKey[KeyCandidate(repoSlug: c.repoSlug, key: c.key), default: []].append(c.sessionID)
-        }
-        var out: [ReconcileBranchMatch] = []
-        for match in matches {
-            guard let sids = sessionsByKey[match.candidate] else { continue }
-            for sid in sids {
-                out.append(ReconcileBranchMatch(
-                    sessionID: sid,
-                    number: match.number,
-                    url: match.url,
-                    state: match.state,
-                    updatedAt: match.updatedAt
-                ))
-            }
-        }
-        return out
-    }
-
-    /// One batched call per backend: GitHub issues a single aliased GraphQL
-    /// query covering every candidate; GitLab issues one REST call per
-    /// (host, candidate) tuple. Returns `nil` on backend error so the
-    /// reconcile pass can skip the cycle without treating a degraded
-    /// response as "no PRs found".
-    private func fetchPRsForReconcile(candidates: [ReconcileCandidate]) async -> [ReconcileBranchMatch]? {
-        guard !candidates.isEmpty else { return [] }
-        let backend = providerManager.codeBackend(for: .github)!
-        do {
-            let matches = try await backend.findRecentPRsForBranches(
-                Self.dedupedBranchCandidates(candidates)
-            )
-            return Self.fanOutMatches(matches, across: candidates)
-        } catch {
-            handleGitHubBackendError(error, operation: "findRecentPRsForBranches(github)")
-            return nil
-        }
-    }
-
-    /// GitLab equivalent: route through the GitLab `CodeBackend` for the given host.
-    private func fetchGitLabMRsForReconcile(
-        candidates: [ReconcileCandidate],
-        host: String
-    ) async -> [ReconcileBranchMatch] {
-        guard !candidates.isEmpty else { return [] }
-        let backend = providerManager.codeBackend(for: .gitlab, host: host)!
-        do {
-            let matches = try await backend.findRecentPRsForBranches(
-                Self.dedupedBranchCandidates(candidates)
-            )
-            return Self.fanOutMatches(matches, across: candidates)
-        } catch {
-            print("[IssueTracker] Reconcile via backend failed for host \(host): \(error.localizedDescription.prefix(200))")
-            return []
-        }
-    }
-
-    /// Project `ReconcileCandidate`s onto the de-duplicated `(repoSlug, branch)`
-    /// pairs the backend needs. Two sessions on the same branch (a duplicated
-    /// session, or reconcile firing before the first session's PR link lands)
-    /// produce a single backend query — we fan the matches back out per
-    /// session in `fanOutMatches`.
-    nonisolated static func dedupedBranchCandidates(_ candidates: [ReconcileCandidate]) -> [BranchCandidate] {
-        var seen: Set<BranchCandidate> = []
-        var out: [BranchCandidate] = []
-        for c in candidates {
-            let bc = BranchCandidate(repoSlug: c.repoSlug, branch: c.branch)
-            if seen.insert(bc).inserted { out.append(bc) }
-        }
-        return out
-    }
-
-    /// Each backend `BranchPRMatch` is duplicated for every `ReconcileCandidate`
-    /// that shares its `(repoSlug, branch)`. This preserves the prior
-    /// per-session sessionID-threading even when two sessions point at the
-    /// same branch — collapsing them via `Dictionary(uniqueKeysWithValues:)`
-    /// would either trap or silently drop one session's PR link.
-    nonisolated static func fanOutMatches(
-        _ matches: [BranchPRMatch],
-        across candidates: [ReconcileCandidate]
-    ) -> [ReconcileBranchMatch] {
-        // Group sessions by their (repoSlug, branch) so a single match maps
-        // to every session that owns that key.
-        var sessionsByBranch: [BranchCandidate: [UUID]] = [:]
-        for c in candidates {
-            let bc = BranchCandidate(repoSlug: c.repoSlug, branch: c.branch)
-            sessionsByBranch[bc, default: []].append(c.sessionID)
-        }
-        var out: [ReconcileBranchMatch] = []
-        for match in matches {
-            guard let sids = sessionsByBranch[match.candidate] else { continue }
-            for sid in sids {
-                out.append(ReconcileBranchMatch(
-                    sessionID: sid,
-                    number: match.number,
-                    url: match.url,
-                    state: match.state,
-                    updatedAt: match.updatedAt
-                ))
-            }
-        }
-        return out
-    }
-
-    /// Persist the reconciliation decisions. Re-checks `appState.links` at
-    /// write time so a concurrent `applySessionPRLinks` or hand-added PR link
-    /// (identified by URL match) wins without leaving a duplicate row.
-    private func applyReconciledPRLinks(_ picks: [ReconcileBranchMatch]) {
-        guard !picks.isEmpty else { return }
-        // Accumulate then persist once — see `applySessionPRLinks` (#304).
-        var newLinks: [SessionLink] = []
-        for pick in picks {
-            let existing = appState.links(for: pick.sessionID)
-            if existing.contains(where: { $0.linkType == .pr || $0.url == pick.url }) { continue }
-            let link = SessionLink(
-                sessionID: pick.sessionID,
-                label: "PR #\(pick.number)",
-                url: pick.url,
-                linkType: .pr
-            )
-            appState.links[pick.sessionID, default: []].append(link)
-            newLinks.append(link)
-        }
-
-        guard !newLinks.isEmpty else { return }
-        // Route through the shared, injected `store` — never a throwaway
-        // `JSONStore()`. A fresh instance reads its own (possibly stale) disk
-        // snapshot and its full-store write can silently clobber a session
-        // another writer just added (#728).
-        store.mutate { data in
-            data.links.append(contentsOf: newLinks)
-        }
-    }
-
-    /// Resolve the org/repo slug (e.g. "corveil/citadel") from a worktree's git remote.
-    private func resolveRepoSlug(worktree: SessionWorktree) -> String {
-        return resolveRepoInfo(worktree: worktree).slug
-    }
-
-    /// Info derived from a worktree's git remote URL: org/repo slug and (for
-    /// GitLab) the host name. Host is empty for github.com remotes.
-    struct RepoInfo: Sendable, Equatable {
-        let slug: String
-        let host: String
-    }
-
-    private func resolveRepoInfo(worktree: SessionWorktree) -> RepoInfo {
-        if let output = try? shellSync(
-            "git", "-C", worktree.repoPath, "remote", "get-url", "origin"
-        ) {
-            var url = output.trimmingCharacters(in: .whitespacesAndNewlines)
-            if url.hasSuffix(".git") { url = String(url.dropLast(4)) }
-            let host = Self.extractHost(fromRemote: url)
-            let slug = Self.extractSlug(fromRemote: url)
-            if !slug.isEmpty {
-                return RepoInfo(slug: slug, host: host)
-            }
-        }
-        if worktree.repoName.contains("/") {
-            return RepoInfo(slug: worktree.repoName, host: "")
-        }
-        return RepoInfo(slug: "", host: "")
-    }
-
-    /// Extract the host ("github.com", "gitlab.example.com") from a git remote URL.
-    /// Handles both SSH (`git@host:org/repo`) and HTTPS (`https://host/org/repo`).
-    /// Returns "" when the URL can't be parsed.
-    nonisolated static func extractHost(fromRemote url: String) -> String {
-        // SSH: git@host:org/repo
-        if let range = url.range(of: #"^[^@]+@([^:]+):"#, options: .regularExpression) {
-            let match = String(url[range])
-            if let at = match.firstIndex(of: "@"), let colon = match.lastIndex(of: ":") {
-                return String(match[match.index(after: at)..<colon])
-            }
-        }
-        // HTTPS: https://host/...
-        if let range = url.range(of: #"^https?://([^/]+)/"#, options: .regularExpression) {
-            let match = String(url[range])
-            let trimmed = match
-                .replacingOccurrences(of: #"^https?://"#, with: "", options: .regularExpression)
-            return trimmed.trimmingCharacters(in: CharacterSet(charactersIn: "/"))
-        }
-        return ""
-    }
-
-    /// Extract the project slug ("org/repo", "group/sub/repo", ...) from a git
-    /// remote URL. Handles both SSH (`git@host:path`) and HTTPS
-    /// (`https://host/path`), and preserves nested-group paths so that GitLab
-    /// projects under nested groups (e.g.
-    /// `big-bang/product/packages/elasticsearch-kibana`) keep their full path.
-    /// Strips a trailing `.git` if present. Returns "" when the URL can't be
-    /// parsed.
-    nonisolated static func extractSlug(fromRemote url: String) -> String {
-        var trimmed = url.trimmingCharacters(in: .whitespacesAndNewlines)
-        if trimmed.hasSuffix(".git") { trimmed = String(trimmed.dropLast(4)) }
-
-        // SSH: git@host:org/repo or user@host:group/sub/repo
-        if let range = trimmed.range(of: #"^[^@/\s]+@[^:/\s]+:"#, options: .regularExpression) {
-            let path = String(trimmed[range.upperBound...])
-            return path.trimmingCharacters(in: CharacterSet(charactersIn: "/"))
-        }
-        // HTTPS: https://host/org/repo
-        if let range = trimmed.range(of: #"^https?://[^/]+/"#, options: .regularExpression) {
-            let path = String(trimmed[range.upperBound...])
-            return path.trimmingCharacters(in: CharacterSet(charactersIn: "/"))
-        }
-        return ""
-    }
-
-    /// Parse the `owner/repo` (or `group/sub/repo`) slug from a PR/MR *web* URL
-    /// such as `https://github.com/owner/repo/pull/123` or
-    /// `https://gitlab.com/group/sub/repo/-/merge_requests/12`. Returns the path
-    /// segments before the `pull` / `merge_requests` / `-` marker, or "" when the
-    /// URL can't be parsed. Distinct from `extractSlug(fromRemote:)`, which
-    /// parses git *remote* URLs (no `/pull/...` suffix).
-    nonisolated static func repoSlug(fromPRURL url: String) -> String {
-        let trimmed = url.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard let range = trimmed.range(of: #"^https?://[^/]+/"#, options: .regularExpression) else {
-            return ""
-        }
-        let path = String(trimmed[range.upperBound...])
-        var segments: [String] = []
-        for segment in path.split(separator: "/").map(String.init) {
-            if segment == "pull" || segment == "merge_requests" || segment == "-" { break }
-            segments.append(segment)
-        }
-        return segments.joined(separator: "/")
-    }
-
-    private func shellSync(_ args: String...) throws -> String {
-        let process = Process()
-        let outPipe = Pipe()
-        let errPipe = Pipe()
-        process.executableURL = URL(fileURLWithPath: "/usr/bin/env")
-        process.arguments = args
-        process.environment = ShellEnvironment.shared.env
-        process.standardOutput = outPipe
-        process.standardError = errPipe
-        try process.run()
-        process.waitUntilExit()
-        let outData = outPipe.fileHandleForReading.readDataToEndOfFile()
-        let errData = errPipe.fileHandleForReading.readDataToEndOfFile()
-        guard process.terminationStatus == 0 else {
-            let stderr = (String(data: errData, encoding: .utf8) ?? "")
-                .trimmingCharacters(in: .whitespacesAndNewlines)
-            let cmd = args.joined(separator: " ")
-            let desc = "`\(cmd)` exited \(process.terminationStatus)"
-                + (stderr.isEmpty ? "" : ": \(stderr)")
-            throw NSError(
-                domain: "IssueTracker",
-                code: Int(process.terminationStatus),
-                userInfo: [NSLocalizedDescriptionKey: desc]
-            )
-        }
-        return String(data: outData, encoding: .utf8) ?? ""
-    }
 
     // MARK: - PR Status (piggyback)
